refactor(dataobj): Translate bloom/regex section predicates to dataset.FuncPredicate
The dataset layer has no concept of a bloom filter or a regex, yet
BloomMatchPredicate and RegexMatchPredicate pushed that domain knowledge
down into it. Define these predicates at the postings section level and
translate them to a dataset.FuncPredicate on pushdown so the matching
logic lives where the domain is understood, keeping the dataset package
agnostic.

Addresses review feedback on #22551:
https://github.com/grafana/loki/pull/22551#discussion_r3467795414
