+++ pkg/dataobj/internal/dataset/predicate.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"iter"
 	"unsafe"
-
-	"github.com/prometheus/prometheus/model/labels"
 )
 
 // Predicate is an expression used to filter rows in a [RowReader].
@@ -74,27 +72,6 @@ type (
 		// If Keep returns true, the row is kept.
 		Keep func(column Column, value Value) bool
 	}
-
-	// BloomMatchPredicate is a [Predicate] which asserts that a row may only be
-	// included if the bloom filter stored in Column's binary value tests positive
-	// for Value.
-	//
-	// Instances of BloomMatchPredicate are ineligible for page filtering: a bloom
-	// blob has no min/max page statistics to prune on.
-	BloomMatchPredicate struct {
-		Column Column
-		Value  []byte
-	}
-
-	// RegexMatchPredicate is a [Predicate] which asserts that a row may only be
-	// included if Column's string value matches Matcher.
-	//
-	// Instances of RegexMatchPredicate are ineligible for page filtering: a regex
-	// has no min/max page statistics to prune on.
-	RegexMatchPredicate struct {
-		Column  Column
-		Matcher *labels.FastRegexMatcher
-	}
 )
 
 func (AndPredicate) isPredicate()         {}
@@ -107,8 +84,6 @@ func (InPredicate) isPredicate()          {}
 func (GreaterThanPredicate) isPredicate() {}
 func (LessThanPredicate) isPredicate()    {}
 func (FuncPredicate) isPredicate()        {}
-func (BloomMatchPredicate) isPredicate()  {}
-func (RegexMatchPredicate) isPredicate()  {}
 
 // WalkPredicate traverses a predicate in depth-first order: it starts by
 // calling fn(p). If fn(p) returns true, WalkPredicate is invoked recursively
@@ -138,8 +113,6 @@ func WalkPredicate(p Predicate, fn func(p Predicate) bool) {
 	case GreaterThanPredicate: // No children.
 	case LessThanPredicate: // No children.
 	case FuncPredicate: // No children.
-	case BloomMatchPredicate: // No children.
-	case RegexMatchPredicate: // No children.
 
 	default:
 		panic(fmt.Sprintf("dataset.WalkPredicate: unsupported predicate type %T", p))
